[compiler] merge aggressive-fusion into generic-fusion
