+++ compiler/lib/Dialect/mhlo/CMakeLists.txt
@@ -94,7 +94,6 @@ add_mlir_dialect_library(ByteIRMhloPasses
   Transforms/FusionOutlining.cpp
   Transforms/MatmulLayoutTransform.cpp
   Transforms/GenericFusion.cpp
-  Transforms/HloAggressiveFusion.cpp
   Transforms/HloFolder.cpp
   Transforms/HloMoveDown.cpp
   Transforms/HloMoveUp.cpp
